+++ cmd/generate_dataset_structs/main.go
@@ -1,6 +1,7 @@
 package main
 
 import (
+	"context"
 	"flag"
 	"fmt"
 	"log"
@@ -13,7 +14,7 @@ import (
 
 var (
 	datasetFlag = flag.String("dataset", "", "dataset name to extract the info from, e.g.: \"microsoft/ms_marco\"")
-	subsetFlag  = flag.String("subset", "", "subset of the dataset to use. This is sometimes used for versioning, e.g.: \"v2.1\". If empty, it will be automatically chosen if only one exists. If multiple subsets exist, lists them and exits.")
+	configFlag  = flag.String("config", "", "config of the dataset to use. This is sometimes used for versioning, e.g.: \"v2.1\". If empty, it will be automatically chosen if only one exists. If multiple configs exist, lists them and exits.")
 	outputFlag  = flag.String("output", "", "file name where to output the generated Go code. If empty, the structs are output to stdout. It requires -package to be set also.")
 	packageFlag = flag.String("package", "", "name of the package to use when outputting a .go file with output")
 )
@@ -31,34 +32,42 @@ func main() {
 
 	log.Printf("Loading dataset %q info...", *datasetFlag)
 	ds := datasets.New(*datasetFlag)
-	info := ds.Info()
-	if info == nil || len(info.DatasetInfo) == 0 {
+	info, err := ds.Info()
+	if err != nil {
+		log.Fatalf("Failed to retrieve dataset info: %+v", err)
+	}
+	if len(info.DatasetInfo) == 0 {
 		log.Fatal("Failed to retrieve dataset info or dataset has no configurations")
 	}
 
-	subset := *subsetFlag
-	if subset == "" {
+	config := *configFlag
+	if config == "" {
 		if len(info.DatasetInfo) == 1 {
-			// Choose the only subset
+			// Choose the only config
 			for k := range info.DatasetInfo {
-				subset = k
+				config = k
 			}
-			log.Printf("Automatically selected the only available subset: %q", subset)
+			log.Printf("Automatically selected the only available config: %q", config)
 		} else {
-			// List available subsets
-			var subsets []string
+			// List available configs
+			var configs []string
 			for k := range info.DatasetInfo {
-				subsets = append(subsets, k)
+				configs = append(configs, k)
 			}
-			sort.Strings(subsets)
-			fmt.Printf("Multiple subsets available: \"%s\"\nPlease specify one using -subset flag.\n", strings.Join(subsets, "\", \""))
+			sort.Strings(configs)
+			fmt.Printf("Multiple configs available: \"%s\"\n", strings.Join(configs, "\", \""))
+			err = ds.DownloadParquetFilesInfo(context.Background(), false)
+			if err != nil {
+				log.Fatalf("Failed to download parquet files info: %+v", err)
+			}
+			fmt.Printf("%s\n", ds)
 			os.Exit(1)
 		}
 	}
 
-	configInfo, ok := info.DatasetInfo[subset]
+	configInfo, ok := info.DatasetInfo[config]
 	if !ok {
-		log.Fatalf("Version %q not found in dataset info.", subset)
+		log.Fatalf("Config %q not found in dataset info.", config)
 	}
 
 	// Generate a root struct name from the dataset name