+++ datasets/info.go
@@ -5,7 +5,6 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"log"
 	"os"
 	"path"
 
@@ -16,21 +15,36 @@ import (
 // Info holds information about a dataset returned by the datasets-server.
 // See the API at: https://datasets-server.huggingface.co/info?dataset=<id>
 type Info struct {
+	// DatasetInfo is keyed on the "config" name.
+	// Some datasets use "config" names to denote versions, e.g. "v1.1".
 	DatasetInfo map[string]ConfigInfo `json:"dataset_info"`
+}
+
+type ParquetFileResponse struct {
+	ParquetFiles []ParquetFile `json:"parquet_files"`
+}
+
+type ParquetFile struct {
+	Dataset  string `json:"dataset"`
+	Config   string `json:"config"`
+	Split    string `json:"split"`
+	URL      string `json:"url"`
+	Filename string `json:"filename"`
+	Size     int64  `json:"size"`
 }
 
 // ConfigInfo holds the info for a specific configuration of a dataset.
 type ConfigInfo struct {
-	Description  string         `json:"description"`
-	Citation     string         `json:"citation"`
-	Homepage     string         `json:"homepage"`
-	License      string         `json:"license"`
+	Description  string             `json:"description"`
+	Citation     string             `json:"citation"`
+	Homepage     string             `json:"homepage"`
+	License      string             `json:"license"`
 	Features     map[string]Feature `json:"features"`
-	BuilderName  string         `json:"builder_name"`
-	ConfigName   string         `json:"config_name"`
-	Version      map[string]any `json:"version"`
-	Splits       map[string]any `json:"splits"`
-	DownloadSize int64          `json:"download_size"`
+	BuilderName  string             `json:"builder_name"`
+	ConfigName   string             `json:"config_name"`
+	Version      map[string]any     `json:"version"`
+	Splits       map[string]Split   `json:"splits"`
+	DownloadSize int64              `json:"download_size"`
 	DatasetSize  int64              `json:"dataset_size"`
 }
 
@@ -45,6 +59,14 @@ type Feature struct {
 
 	// Used by features like ClassLabel
 	Names []string `json:"names,omitempty"`
+}
+
+// Split holds information about a dataset split (e.g. "train", "test", "validation").
+type Split struct {
+	DatasetName string `json:"dataset_name,omitempty"`
+	Name        string `json:"name"`
+	NumBytes    int64  `json:"num_bytes"`
+	NumExamples int64  `json:"num_examples"`
 }
 
 // UnmarshalJSON implements custom JSON unmarshaling to handle dynamic "feature" values.
@@ -123,14 +145,14 @@ func (c *ConfigInfo) String() string {
 // If it hasn't been downloaded or loaded from the cache yet, it loads it first.
 // It may return nil if there was an issue with the downloading of the Info json from HuggingFace.
 // Try DownloadDatasetInfo to get an error.
-func (d *Dataset) Info() *Info {
+func (d *Dataset) Info() (*Info, error) {
 	if d.info == nil {
 		err := d.DownloadDatasetInfo(context.Background(), false)
 		if err != nil {
-			log.Printf("Error while downloading info about Dataset: %+v", err)
+			return nil, errors.WithMessagef(err, "failed to download info for dataset %q", d.ID)
 		}
 	}
-	return d.info
+	return d.info, nil
 }
 
 // datasetServerURL for the API that returns the info about a dataset.
@@ -174,5 +196,58 @@ func (d *Dataset) DownloadDatasetInfo(ctx context.Context, forceDownload bool) e
 		return errors.Wrapf(err, "failed to parse info for dataset in %q (downloaded from %q)", infoFilePath, d.datasetServerURL())
 	}
 	d.info = newInfo
+	return nil
+}
+
+// datasetServerParquetURL returns the URL to query parquet files info.
+func (d *Dataset) datasetServerParquetURL() string {
+	return fmt.Sprintf("https://datasets-server.huggingface.co/parquet?dataset=%s", d.ID)
+}
+
+// GetParquetFiles returns the list of parquet files for the dataset.
+// If it hasn't been downloaded or loaded from the cache yet, it loads it first.
+func (d *Dataset) GetParquetFiles() ([]ParquetFile, error) {
+	if d.Files == nil {
+		err := d.DownloadParquetFilesInfo(context.Background(), false)
+		if err != nil {
+			return nil, errors.WithMessagef(err, "failed to download parquet files info for dataset %q", d.ID)
+		}
+	}
+	return d.Files, nil
+}
+
+// DownloadParquetFilesInfo downloads the parquet info using the datasets-server.
+//
+// If forceDownload is set to true, it ignores the cached one.
+func (d *Dataset) DownloadParquetFilesInfo(ctx context.Context, forceDownload bool) error {
+	if d.Files != nil && !forceDownload {
+		return nil
+	}
+
+	infoFilePath, err := d.Repo.CacheDir()
+	if err != nil {
+		return err
+	}
+	infoFilePath = path.Join(infoFilePath, "dataset_parquet.json")
+
+	// Download info file if needed.
+	if !files.Exists(infoFilePath) || forceDownload {
+		err := d.Repo.GetDownloadManager().LockedDownload(ctx, d.datasetServerParquetURL(), infoFilePath, forceDownload, nil)
+		if err != nil {
+			return errors.WithMessagef(err, "failed to download dataset parquet files info")
+		}
+	}
+
+	infoJson, err := os.ReadFile(infoFilePath)
+	if err != nil {
+		return errors.Wrapf(err, "failed to read parquet info for dataset from disk in %q -- remove the file if you want to have it re-downloaded", infoFilePath)
+	}
+
+	decoder := json.NewDecoder(bytes.NewReader(infoJson))
+	var response ParquetFileResponse
+	if err = decoder.Decode(&response); err != nil {
+		return errors.Wrapf(err, "failed to parse parquet info for dataset in %q (downloaded from %q)", infoFilePath, d.datasetServerParquetURL())
+	}
+	d.Files = response.ParquetFiles
 	return nil
 }